pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ramil Albakov <ramil.albakov2@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
gen-diff = "gendiff.scripts.gen_diff:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
ipython = "^8.8.0"
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
